vps-model-gateway/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
boto3==1.34.56
PyYAML==6.0.1
prometheus-fastapi-instrumentator==7.0.0
redis-py-cluster==2.1.3
